move engineDataConfig to controller/model
